Move css location and fix linter errors
