--- src/components/Details.module.css
@@ -1,83 +1,0 @@
-.detimg {
-width: 70%;
-    height: 100%;
-    border-radius: 14px;
-}
-
-.details {
-display: flex;
-    justify-content: space-between;
-    width: 84vw;
-    height: 100vh;
-    padding: 20px;
-    gap: 30px;
-}
-
-.hoteltextdetails{
-      display: flex;
-  justify-content: flex-start;
-  flex-direction: column;
-  gap: 20px;
-  width: 40%;
-}
-
-.hoteldetailsname{
-      font-size: 26px;
-  font-weight: 600;
-  padding: 0 10px;
-}
-.hoteldetailsdescription{
-  font-weight: 500;
-  color: #727171;
-  font-size: 14px;
-  padding: 10px;
-  background-color: #f1f1f1;
-  border-radius: 4px;
-}
-
-.durationcontainer,
-.hotelpricecontainer{
-      display: flex;
-  justify-content: space-between;
-  padding: 10px;
-  color: #727171;
-  font-size: 14px;
-  font-weight: 500;
-}
-
-.durationcontainer {
-  background-color: #f1f1f1;
-  padding: 10px;
-  margin-bottom: 50px;
-  border-radius: 4px;
-}
-
-.textheading {
-  font-weight: 600;
-  font-weight: 13px;
-}
-
-.morehotelslink {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 5px;
-  padding: 10px;
-  background-color: #f1f1f1;
-  font-size: 14px;
-  font-weight: 500;
-  color: #727171;
-  border-radius: 4px;
-}
-
-.reserve{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: green;
-    color: white;
-    border-radius: 4px;
-    font-size: 20px;
-    width: 100px;
-    height: 40px;
-}
