--- src/components/Home.module.css
@@ -1,131 +1,0 @@
-.hoteldiv {
-    display: flex;
-    margin-top: 3%;
-}
-
-.hotelimg {
-    width: 50% !important;
-    height: 300px;
-    border-radius: 50%;
-}
-
-a {
-    text-decoration: none;
-    color: black;
-}
-
-.hotelDesc {
-    margin-top: 3%;
-    color: gray;
-}
-
-.homediv {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 100vh;
-}
-
-.hotelinfo {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.hotelname{
-    margin-top: 3%;
-    color: gray;
-}
-
-.hoteltitle {
-    margin-top: 8%;
-    color: gray;
-}
-
-.hotelsubtitle{
-    margin-top: 3%;
-    color: gray;
-}
-
-.hotelicons{
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 20px;
-    margin-top: 3%;
-}
-
-.hotelicon{
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 20px;
-    color: #908e8e;
-    border: 3px solid #908e8e;
-    padding: 2px;
-    border-radius: 50%;
-}
-
-.dotedbordercontainer{
-    border-bottom: 4px dotted #d3d3d3;
-    margin-top: 3%;
-        width: 100px;
-        align-self: center;
-}
-
-.carouselButton {
-    background: #0ab86d;
-    border: none;
-    color: #333;
-    padding: 10px;
-    text-align: center;
-    display: inline-block;
-    margin: 4px 2px;
-    cursor: pointer;
-    border-radius: 50px 50px 5px 5px;
-    font-size: 16px;
-    position: absolute;
-    top: 30%;
-    transform: translateY(-50%);
-    z-index: 2;
-        width: 60px;
-            height: 50px;
-}
-
-.carouselButton.prev {
-    left: 1px;
-        border-radius: 5px 50px 50px 5px;
-}
-
-.carouselButton.next {
-    right: 1px;
-        border-radius: 50px 5px 5px 50px;
-}
-
-.svgicon{
-    color: white;
-}
-
-@media screen and (min-width: 769px) {
-    .hotelname {
-            margin-top: 2%;
-        }
-    
-        .hoteltitle {
-            margin-top: 6%;
-        }
-    
-        .hotelsubtitle {
-            margin-top: 2%;
-        }
-                .hotelicons {
-                    margin-top: 2%;
-                }
-                                .dotedbordercontainer {
-                                    margin-top: 2%;
-                                }
-
-                                                                .hotelimg {
-                                                                    width: 70% !important;
-                                                                }
-}
