+++ src/styles/Authentication.css
@@ -28,3 +28,15 @@ input {
   background-color: transparent;
   border: 1px solid green;
 }
+
+.logout {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #b4c424;
+  color: white;
+  border-radius: 4px;
+  font-size: 15px;
+  width: 100px;
+  height: 32px;
+}
