+++ src/styles/Details.module.css
@@ -1,0 +1,120 @@
+.detimg {
+  width: 100%;
+  height: 100%;
+  border-radius: 14px;
+}
+
+.details {
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+  height: 100vh;
+  padding: 20px;
+  gap: 30px;
+}
+
+.hoteltextdetails {
+  display: flex;
+  justify-content: space-between;
+  gap: 20px;
+}
+
+.hoteldetailsname {
+  font-size: 26px;
+  font-weight: 600;
+  padding: 0 10px;
+}
+
+.hoteldetailsdescription {
+  font-weight: 500;
+  color: #727171;
+  font-size: 14px;
+  padding: 10px;
+  background-color: #f1f1f1;
+  border-radius: 4px;
+}
+
+.durationcontainer,
+.hotelpricecontainer {
+  display: flex;
+  justify-content: space-between;
+  padding: 10px;
+  color: #727171;
+  font-size: 14px;
+  font-weight: 500;
+}
+
+.durationcontainer {
+  background-color: #f1f1f1;
+  padding: 10px;
+  margin-bottom: 50px;
+  border-radius: 4px;
+}
+
+.textheading {
+  font-weight: 600;
+  font-size: 13px;
+}
+
+.morehotelslink {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  gap: 5px;
+  padding: 10px;
+  background-color: #f1f1f1;
+  font-size: 14px;
+  font-weight: 500;
+  color: #727171;
+  border-radius: 4px;
+  margin-top: 21%;
+}
+
+.reserve {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #b4c424;
+  color: white;
+  border-radius: 4px;
+  font-size: 15px;
+  width: 100px;
+  height: 32px;
+}
+
+.links {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.detailer {
+  display: flex;
+  flex-direction: column;
+  width: 50%;
+  gap: 10px;
+}
+
+@media screen and (min-width: 769px) {
+  .details {
+    flex-direction: row;
+    width: 84vw;
+    height: 100vh;
+    padding: 20px;
+    gap: 30px;
+  }
+
+  .hoteltextdetails {
+    flex-direction: column;
+    width: 40%;
+    justify-content: flex-start;
+  }
+
+  .detimg {
+    width: 70%;
+  }
+
+  .detailer {
+    width: 100%;
+  }
+}
